+++ internal/virtualterminal/main.go
@@ -1,0 +1,5 @@
+package virtualterminal
+
+func Enable(handle int) (func(), error) {
+	return enable(handle)
+}