+++ internal/virtualterminal/main_unix.go
@@ -1,7 +1,7 @@
 //go:build !windows
 // +build !windows
 
-package tty10pe
+package virtualterminal
 
 func enable(handle int) (func(), error) {
 	return func() {}, nil