+++ internal/winch/main_unix.go
@@ -1,0 +1,44 @@
+//go:build !windows
+
+package winch
+
+import (
+	"context"
+	"os"
+	"os/signal"
+	"sync"
+
+	"golang.org/x/sys/unix"
+)
+
+func notice(onResize func(int, int)) (func(), error) {
+	ctx, cancel := context.WithCancel(context.Background())
+	var wg sync.WaitGroup
+
+	ch := make(chan os.Signal, 1)
+	signal.Notify(ch, unix.SIGWINCH)
+	println("a")
+
+	wg.Add(1)
+	go func() {
+		for {
+			select {
+			case <-ctx.Done():
+				signal.Stop(ch)
+				close(ch)
+				wg.Done()
+				return
+			case <-ch:
+				w, h, err := getSize()
+				if err == nil {
+					onResize(w, h)
+				}
+			}
+		}
+	}()
+
+	return func() {
+		cancel()
+		wg.Wait()
+	}, nil
+}